+++ bot/src/main/java/de/chojo/gamejam/commands/team/Create.java
@@ -1,7 +1,6 @@
 package de.chojo.gamejam.commands.team;
 
 
-import de.chojo.gamejam.commands.TeamSubCommand;
 import de.chojo.gamejam.data.TeamData;
 import de.chojo.gamejam.data.wrapper.jam.Jam;
 import de.chojo.gamejam.data.wrapper.team.JamTeam;
